Moved to util.nio package, changes to comments
