+++ src/main/java/org/broadinstitute/hellbender/utils/Utils.java
@@ -1142,127 +1142,6 @@ public final class Utils {
         return sorted.get(sorted.size() / 2);
     }
 
-    public static class LineIterator implements AutoCloseable, Iterable<String> {
-        private final Stream<String> lines;
 
-        /**
-         * Returns an iterator so you can iterate over the lines in the text file like so:
-         * for (String line: new Utils.LineIterator(path)) {
-         *   // do something with the line
-         * }
-         *
-         * It's also closeable so you can close it when done, or use it in a try-with-resources
-         * to close it automatically.
-         *
-         * @param path path to a text file.
-         * @throws UserException if we cannot open the file for reading.
-         */
-        public LineIterator(Path path) throws UserException {
-            try {
-                lines = Files.lines(path);
-            } catch (IOException x) {
-                throw new UserException("Error reading " + path.toString() + ": " + x.getMessage());
-            }
-        }
-
-        /**
-         * Closes this resource, relinquishing any underlying resources.
-         * This method is invoked automatically on objects managed by the
-         * {@code try}-with-resources statement.
-         *
-         * <p>While this interface method is declared to throw {@code
-         * Exception}, implementers are <em>strongly</em> encouraged to
-         * declare concrete implementations of the {@code close} method to
-         * throw more specific exceptions, or to throw no exception at all
-         * if the close operation cannot fail.
-         *
-         * <p> Cases where the close operation may fail require careful
-         * attention by implementers. It is strongly advised to relinquish
-         * the underlying resources and to internally <em>mark</em> the
-         * resource as closed, prior to throwing the exception. The {@code
-         * close} method is unlikely to be invoked more than once and so
-         * this ensures that the resources are released in a timely manner.
-         * Furthermore it reduces problems that could arise when the resource
-         * wraps, or is wrapped, by another resource.
-         *
-         * <p><em>Implementers of this interface are also strongly advised
-         * to not have the {@code close} method throw {@link
-         * InterruptedException}.</em>
-         *
-         * This exception interacts with a thread's interrupted status,
-         * and runtime misbehavior is likely to occur if an {@code
-         * InterruptedException} is {@linkplain Throwable#addSuppressed
-         * suppressed}.
-         *
-         * More generally, if it would cause problems for an
-         * exception to be suppressed, the {@code AutoCloseable.close}
-         * method should not throw it.
-         *
-         * <p>Note that unlike the {@link Closeable#close close}
-         * method of {@link Closeable}, this {@code close} method
-         * is <em>not</em> required to be idempotent.  In other words,
-         * calling this {@code close} method more than once may have some
-         * visible side effect, unlike {@code Closeable.close} which is
-         * required to have no effect if called more than once.
-         *
-         * However, implementers of this interface are strongly encouraged
-         * to make their {@code close} methods idempotent.
-         *
-         * @throws Exception if this resource cannot be closed
-         */
-        @Override
-        public void close() {
-            lines.close();
-        }
-
-        /**
-         * @return an Iterator over the lines (as Strings).
-         */
-        @Override
-        public Iterator<String> iterator() {
-            return lines.iterator();
-        }
-
-        /**
-         * Performs the given action for each element of the {@code Iterable}
-         * until all elements have been processed or the action throws an
-         * exception.  Unless otherwise specified by the implementing class,
-         * actions are performed in the order of iteration (if an iteration order
-         * is specified).  Exceptions thrown by the action are relayed to the
-         * caller.
-         *
-         * @param action The action to be performed for each element
-         * @throws NullPointerException if the specified action is null
-         * @implSpec <p>The default implementation behaves as if:
-         * <pre>{@code
-         *     for (T t : this)
-         *         action.accept(t);
-         * }</pre>
-         * @since 1.8
-         */
-        @Override
-        public void forEach(Consumer<? super String> action) {
-            lines.forEach(action);
-        }
-
-        /**
-         * Creates a {@link Spliterator} over the elements described by this
-         * {@code Iterable}.
-         *
-         * @return a {@code Spliterator} over the elements described by this {@code Iterable}.
-         * @implSpec The default implementation creates an <em><a href="Spliterator.html#binding">early-binding</a></em>
-         * spliterator from the iterable's {@code Iterator}.  The spliterator inherits the
-         * <em>fail-fast</em> properties of the iterable's iterator.
-         * @implNote The default implementation should usually be overridden.  The spliterator returned by
-         * the default implementation has poor splitting capabilities, is unsized, and does not report any
-         * spliterator characteristics. Implementing classes can nearly always provide a better
-         * implementation.
-         * @since 1.8
-         */
-        @Override
-        public Spliterator<String> spliterator() {
-            return lines.spliterator();
-        }
-    }
 
 }
